Renames app module to app_builder.
